refactor(parser): address review feedback
- Split rule parser tests into submodules to keep files under 400 lines.

- Use PatternTokenStream::push_error() to avoid direct error field access.

- Import and call span_utils::shift_errors directly (remove wrapper).
