+++ src/parser/tests/rules/mod.rs
@@ -1,0 +1,9 @@
+//! Rule parsing tests.
+//!
+//! Ensures rule heads, bodies, and error cases are handled.
+
+mod aggregations;
+mod body_terms;
+mod invalid;
+mod parsing;
+mod split_assignment;
